Branch pijus

+++ autominus.client/src/API/DeleteCar.jsx
@@ -1,17 +1,22 @@
-async function DeleteCar(carID) {
-    try {
-        const response = await fetch(`car/${carID}`, {
-            method: "DELETE",
-        });
+async function DeleteCar(carID, carUserId) {
+    const userId = localStorage.getItem("userId");
+    if (userId == carUserId) {
+        try {
+            const response = await fetch(`car/${carID}`, {
+                method: "DELETE",
+            });
 
-        if (!response.ok) {
-            throw new Error(`HTTP error! Status: ${response.status}`);
+            if (!response.ok) {
+                throw new Error(`HTTP error! Status: ${response.status}`);
+            }
+
+        } catch (error) {
+            console.error("Error deleting car:", error);
+            throw error;
         }
-
-    } catch (error) {
-        console.error("Error deleting car:", error);
-        throw error;
     }
+  //  console.error("Error deleting car:", error);
+   //throw error;
 }
 
 export default DeleteCar;
